+++ Changes.md
@@ -1,3 +1,4 @@
+* Automatically select short practice round columns when loading a schedule
 * Allow one to save a modified schedule from the SchedulerUI
 * #1096 - Allow one to optimize tables in groups
 * #1102 - Send to web sockets in the background to speed up web clients